Renaming unit test class
